requirements.txt
Django==4.2.4
djangorestframework==3.14.0
requests==2.31.0
aiogram==2.9.2
torch==2.0.1
cdifflib==1.2.6
numpy==1.25.2
pandas==2.0.3